Skype 4.3.0.37 FreeBSD port
Comment some things out in the Makefile that allow building and
installation without going through the list of dependencies and trying
to install the f10- versions.

Tested to work with linux-f20 libs. Didn't attempt a skype call, just
added a contact and tested that chatting works.

This is a rip-off port based on what netchild had put together, for
testing the linux-f20 libs.

Signed-off-by: Vassilis Laganakos <vassilis@einval.com>